create multiselect component
